Extend EmacsVterm.jl interface to be future-proof
The idea is that the code in EmacsVterm.jl can specify which kind of
data it is sending to Emacs and Emacs will decide how to display the
data.

Currently, the only supported kind of data is "documentation", which
is shown in the *julia-doc* buffer, but later, things like inline
display of images or rendered latex code might be added. If the user
has a newer version of EmacsVterm.jl trying to send data not supported
by user's version of julia-repl, an error message is shown.